pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradebot"
version = "0.1.0"
description = "Screenshot-aided trading bot with simulator and Kraken client"
authors = [{name = "You"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "pytesseract>=0.3.10",
    "Pillow>=10.3",
    "Typer>=0.12",
    "rich>=13.7",
    "tenacity>=8.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"tradebot" = ["py.typed"]

[tool.black]
line-length = 100
